CIRC-717: Fix tests.

+++ src/test/java/api/support/fixtures/TemplateContextMatchers.java
@@ -229,8 +229,8 @@ public class TemplateContextMatchers {
       hasJsonPath("fee.type", is(account.getFeeFineType())),
       hasJsonPath("fee.actionType", is(action.getTypeAction())),
       hasJsonPath("fee.actionDateTime", isEquivalentTo(action.getDateAction())),
-      hasJsonPath("fee.actionAmount", is(action.getAmountAction())),
-      hasJsonPath("fee.amount", is(account.getAmount())),
+      hasJsonPath("fee.actionAmount", is(action.getAmountAction().doubleValue())),
+      hasJsonPath("fee.amount", is(account.getAmount().doubleValue())),
       hasJsonPath("fee.balance", is(action.getBalance()))
     );
   }
